Bug 1834870: Move sync telemetry into application services
