+++ android-components/components/browser/storage-sync/src/main/java/mozilla/components/browser/storage/sync/Connection.kt
@@ -8,8 +8,8 @@ import androidx.annotation.GuardedBy
 import mozilla.appservices.places.PlacesApi
 import mozilla.appservices.places.PlacesReaderConnection
 import mozilla.appservices.places.PlacesWriterConnection
+import mozilla.appservices.syncmanager.SyncTelemetry
 import mozilla.components.concept.sync.SyncAuthInfo
-import mozilla.components.support.sync.telemetry.SyncTelemetry
 import java.io.Closeable
 import java.io.File
