+++ android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/FxaDeviceConstellation.kt
@@ -11,6 +11,7 @@ import androidx.lifecycle.LifecycleOwner
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.withContext
 import mozilla.appservices.fxaclient.FxaException
+import mozilla.appservices.syncmanager.SyncTelemetry
 import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.concept.sync.AccountEvent
 import mozilla.components.concept.sync.AccountEventsObserver
@@ -26,7 +27,6 @@ import mozilla.components.concept.sync.ServiceResult
 import mozilla.components.support.base.log.logger.Logger
 import mozilla.components.support.base.observer.Observable
 import mozilla.components.support.base.observer.ObserverRegistry
-import mozilla.components.support.sync.telemetry.SyncTelemetry
 import mozilla.appservices.fxaclient.PersistedFirefoxAccount as FirefoxAccount
 
 internal sealed class FxaDeviceConstellationException : Exception() {
@@ -167,7 +167,10 @@ class FxaDeviceConstellation(
             when (outgoingCommand) {
                 is DeviceCommandOutgoing.SendTab -> {
                     account.sendSingleTab(targetDeviceId, outgoingCommand.title, outgoingCommand.url)
-                    SyncTelemetry.processFxaTelemetry(account.gatherTelemetry(), crashReporter)
+                    val errors: List<Throwable> = SyncTelemetry.processFxaTelemetry(account.gatherTelemetry())
+                    for (error in errors) {
+                        crashReporter?.submitCaughtException(error)
+                    }
                 }
                 else -> logger.debug("Skipped sending unsupported command type: $outgoingCommand")
             }
@@ -203,7 +206,10 @@ class FxaDeviceConstellation(
             false
         } else {
             processEvents(events)
-            SyncTelemetry.processFxaTelemetry(account.gatherTelemetry(), crashReporter)
+            val errors: List<Throwable> = SyncTelemetry.processFxaTelemetry(account.gatherTelemetry())
+            for (error in errors) {
+                crashReporter?.submitCaughtException(error)
+            }
             true
         }
     }
