--- android-components/components/support/sync-telemetry/build.gradle
@@ -1,67 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-buildscript {
-    repositories {
-        maven {
-            url "https://maven.mozilla.org/maven2"
-        }
-
-        dependencies {
-            classpath "org.mozilla.telemetry:glean-gradle-plugin:${Versions.mozilla_glean}"
-        }
-    }
-}
-
-plugins {
-    id "com.jetbrains.python.envs" version "$python_envs_plugin"
-}
-
-apply plugin: 'com.android.library'
-apply plugin: 'kotlin-android'
-
-android {
-    compileSdkVersion config.compileSdkVersion
-
-    defaultConfig {
-        minSdkVersion config.minSdkVersion
-        targetSdkVersion config.targetSdkVersion
-
-        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
-    }
-
-    buildTypes {
-        release {
-            minifyEnabled false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-
-    namespace 'mozilla.components.support.sync.telemetry'
-}
-
-dependencies {
-    implementation project(':support-base')
-
-    implementation project(':service-glean')
-
-    implementation ComponentsDependencies.kotlin_coroutines
-    implementation ComponentsDependencies.mozilla_sync15
-
-    testImplementation ComponentsDependencies.testing_robolectric
-    testImplementation project(":support-test")
-    testImplementation ComponentsDependencies.androidx_test_core
-    testImplementation ComponentsDependencies.androidx_test_junit
-    testImplementation ComponentsDependencies.androidx_work_testing
-
-    testImplementation ComponentsDependencies.testing_mockito
-    testImplementation ComponentsDependencies.mozilla_glean_forUnitTests
-}
-
-apply from: '../../../android-lint.gradle'
-apply from: '../../../publish.gradle'
-ext.configurePublish(config.componentsGroupId, archivesBaseName, project.ext.description)
-
-ext.gleanGenerateMarkdownDocs = true
-apply plugin: "org.mozilla.telemetry.glean-gradle-plugin"
