+++ focus-android/.buildconfig.yml
@@ -59,7 +59,6 @@ projects:
     - support-locale
     - support-rusthttp
     - support-rustlog
-    - support-sync-telemetry
     - support-test
     - support-test-libstate
     - support-utils
